+++ Sources/DockAppToggler/Core/Extensions/NSAppearance+Extensions.swift
@@ -1,0 +1,7 @@
+import AppKit
+
+extension NSAppearance {
+    var isDarkMode: Bool {
+        self.bestMatch(from: [.darkAqua, .aqua]) == .darkAqua
+    }
+}
