+++ Sources/DockAppToggler/Features/Dock/DockWatcher.swift
