+++ Sources/DockAppToggler/Features/Updates/UpdateController.swift
@@ -1,0 +1,56 @@
+import AppKit
+import Sparkle
+import UserNotifications
+
+@MainActor
+class UpdateController: NSObject, SPUStandardUserDriverDelegate {
+    nonisolated var supportsGentleScheduledUpdateReminders: Bool {
+        return true
+    }
+    
+    nonisolated func standardUserDriverWillHandleShowingUpdate(_ handleShowingUpdate: Bool, forUpdate update: SUAppcastItem, state: SPUUserUpdateState) {
+        // Capture the values we need before starting the task
+        let version = update.displayVersionString
+        let isUserInitiated = state.userInitiated
+        
+        Task { @MainActor in
+            // When an update alert will be presented, place the app in the foreground
+            NSApp.setActivationPolicy(.regular)
+            
+            if !isUserInitiated {
+                // Add a badge to the app's dock icon indicating one alert occurred
+                NSApp.dockTile.badgeLabel = "1"
+                
+                // Post a user notification
+                let content = UNMutableNotificationContent()
+                content.title = "A new update is available"
+                content.body = "Version \(version) is now available"
+                
+                let request = UNNotificationRequest(identifier: "UpdateCheck", content: content, trigger: nil)
+                
+                do {
+                    try await UNUserNotificationCenter.current().add(request)
+                } catch {
+                    Logger.error("Failed to add notification: \(error)")
+                }
+            }
+        }
+    }
+    
+    nonisolated func standardUserDriverDidReceiveUserAttention(forUpdate update: SUAppcastItem) {
+        Task { @MainActor in
+            // Clear the dock badge indicator for the update
+            NSApp.dockTile.badgeLabel = ""
+            
+            // Dismiss active update notifications without await since it's not async
+            UNUserNotificationCenter.current().removeDeliveredNotifications(withIdentifiers: ["UpdateCheck"])
+        }
+    }
+    
+    nonisolated func standardUserDriverWillFinishUpdateSession() {
+        Task { @MainActor in
+            // Put app back in background when the user session for the update finished
+            NSApp.setActivationPolicy(.accessory)
+        }
+    }
+}
