+++ Sources/DockAppToggler/UI/Components/CloseButton.swift
