+++ Sources/DockAppToggler/UI/Components/MinimizeButton.swift
