+++ Sources/DockAppToggler/UI/Windows/WindowChooserController.swift
