+++ Sources/DockAppToggler/UI/Windows/WindowChooserView.swift
